test: update tests for dialect-native type markers
Replace all references to removed core type markers (Int, Text, Bool,
etc.) with dialect-native types (sqlite::types::Integer, postgres::
types::Int4, etc.) in test sources and regenerate stderr expectations.

+++ tests/ui/aggregate_types/fail/sum_int_wrong_width.rs
@@ -16,7 +16,7 @@ struct Schema {
 
 #[derive(SQLiteFromRow)]
 struct SumRow {
-    total: Option<i64>,
+    total: Option<f64>,
 }
 
 fn main() {
